move _define_layout() to BaseCalculatorWindow
